+++ coreexample/src/main/kotlin/me/proton/android/core/coreexample/di/PaymentsModule.kt
@@ -34,9 +34,10 @@ object PaymentsModule {
 
     @Provides
     @Singleton
-    fun providePaymentsRepository(apiProvider: ApiProvider): PaymentsRepository =
-        PaymentsRepositoryImpl(apiProvider)
-
+    fun providePaymentsRepository(
+        apiProvider: ApiProvider
+    ): PaymentsRepository =
+        PaymentsRepositoryImpl(provider = apiProvider)
 
     @Provides
     @Singleton