+++ payment/data/src/main/kotlin/me/proton/core/payment/data/api/response/PlanResponse.kt
@@ -42,6 +42,8 @@ internal data class PlanResponse(
     val maxDomains: Int,
     @SerialName("MaxAddresses")
     val maxAddresses: Int,
+    @SerialName("MaxCalendars")
+    val maxCalendars: Int,
     @SerialName("MaxSpace")
     val maxSpace: Long,
     @SerialName("MaxMembers")
@@ -58,8 +60,22 @@ internal data class PlanResponse(
     val maxTier: Int
 ) {
     fun toPlan(): Plan = Plan(
-        id = id, type = type, cycle = cycle, name = name, title = title, currency = currency, amount = amount,
-        maxDomains = maxDomains, maxAddresses = maxAddresses, maxSpace = maxSpace, maxMembers = maxMembers,
-        maxVPN = maxVPN, services = services, features = features, quantity = quantity, maxTier = maxTier
+        id = id,
+        type = type,
+        cycle = cycle,
+        name = name,
+        title = title,
+        currency = currency,
+        amount = amount,
+        maxDomains = maxDomains,
+        maxAddresses = maxAddresses,
+        maxCalendars = maxCalendars,
+        maxSpace = maxSpace,
+        maxMembers = maxMembers,
+        maxVPN = maxVPN,
+        services = services,
+        features = features,
+        quantity = quantity,
+        maxTier = maxTier
     )
 }