+++ payment/domain/src/main/kotlin/me/proton/core/payment/domain/Const.kt
@@ -16,18 +16,7 @@
  * along with ProtonCore.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.proton.core.plan.presentation.entity
+package me.proton.core.payment.domain
 
-import me.proton.core.presentation.utils.Price
-import me.proton.core.util.kotlin.exhaustive
-
-enum class Cycle {
-    MONTHLY, YEARLY;
-
-    fun getPrice(pricing: PlanPricing): Price {
-        return when (this) {
-            MONTHLY -> pricing.monthly
-            YEARLY -> pricing.yearly
-        }.toDouble().exhaustive
-    }
-}
+// the maximum number of plan items that could be purchased from a mobile app
+const val MAX_PLAN_QUANTITY = 1