+++ payment/presentation/src/main/kotlin/me/proton/core/payment/presentation/viewmodel/PaymentOptionsViewModel.kt
@@ -56,12 +56,12 @@ class PaymentOptionsViewModel @Inject constructor(
     @ApplicationContext private val context: Context,
     private val availablePaymentMethods: GetAvailablePaymentMethods,
     private val getCurrentSubscription: GetCurrentSubscription,
-    private val validatePlanSubscription: ValidateSubscriptionPlan,
-    private val createPaymentTokenWithNewCreditCard: CreatePaymentTokenWithNewCreditCard,
-    private val createPaymentTokenWithNewPayPal: CreatePaymentTokenWithNewPayPal,
-    private val createPaymentTokenWithExistingPaymentMethod: CreatePaymentTokenWithExistingPaymentMethod,
-    private val performSubscribe: PerformSubscribe,
-    private val getCountry: GetCountry
+    validatePlanSubscription: ValidateSubscriptionPlan,
+    createPaymentTokenWithNewCreditCard: CreatePaymentTokenWithNewCreditCard,
+    createPaymentTokenWithNewPayPal: CreatePaymentTokenWithNewPayPal,
+    createPaymentTokenWithExistingPaymentMethod: CreatePaymentTokenWithExistingPaymentMethod,
+    performSubscribe: PerformSubscribe,
+    getCountry: GetCountry
 ) : BillingViewModel(
     validatePlanSubscription,
     createPaymentTokenWithNewCreditCard,
@@ -102,7 +102,7 @@ class PaymentOptionsViewModel @Inject constructor(
         val currentSubscription = getCurrentSubscription(userId)
         currentSubscription?.let {
             it.plans.forEach { plan ->
-                currentPlans.add(plan.id)
+                currentPlans.add(plan.name)
             }
         } ?: run {
             emit(State.Error.SubscriptionInRecoverableError)
@@ -145,34 +145,34 @@ class PaymentOptionsViewModel @Inject constructor(
 
     fun subscribe(
         userId: UserId?,
-        planId: String,
+        planName: String, // plan name
         codes: List<String>? = null,
         currency: Currency,
         cycle: SubscriptionCycle,
         paymentType: PaymentType
     ) = subscribe(
-        userId, currentPlans.plus(planId), codes, currency, cycle, paymentType
+        userId, currentPlans.plus(planName), codes, currency, cycle, paymentType
     )
 
     fun onThreeDSTokenApproved(
         userId: UserId?,
-        planId: String,
+        planName: String,
         codes: List<String>? = null,
         amount: Long,
         currency: Currency,
         cycle: SubscriptionCycle,
         token: String
     ) = onThreeDSTokenApproved(
-        userId, currentPlans.plus(planId), codes, amount, currency, cycle, token
+        userId, currentPlans.plus(planName), codes, amount, currency, cycle, token
     )
 
     fun validatePlan(
         userId: UserId?,
-        planId: String,
+        planName: String,
         codes: List<String>? = null,
         currency: Currency,
         cycle: SubscriptionCycle
-    ) = validatePlan(userId, currentPlans.plus(planId).distinct(), codes, currency, cycle)
+    ) = validatePlan(userId, currentPlans.plus(planName).distinct(), codes, currency, cycle)
 
     companion object {
         const val NO_ACTIVE_SUBSCRIPTION = 22110